% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzCountSignals.R
\name{as.list.CountSignals}
\alias{as.list.CountSignals}
\title{Converts the container to a list \code{l} such that
\code{l[[i]]} is the i-th signal.}
\usage{
\method{as.list}{CountSignals}(x, ...)
}
\arguments{
\item{x}{A CountSignals object}

\item{...}{not used}
}
\value{
a list \code{l} such that
\code{l[[i]]} is \code{x[i]}.
}
\description{
Converts the container to a list \code{l} such that
\code{l[[i]]} is the i-th signal.
}

