% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzzCountSignals.R
\name{as.list.CountSignals}
\alias{as.list.CountSignals}
\title{Converts the container to a list \code{l} such that
\code{l[[i]]} is the i-th signal.}
\usage{
as.list.CountSignals(x, ...)
}
\description{
Converts the container to a list \code{l} such that
\code{l[[i]]} is the i-th signal.
}

